requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==5.0
django-filter==23.5
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
environ==1.0
gunicorn==21.2.0
inflection==0.5.1
packaging==23.2
psycopg2==2.9.9
PyJWT==2.8.0
python-environ==0.4.54
pytz==2023.3.post1
PyYAML==6.0.1
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
whitenoise==6.6.0
django-cors-headers==3.10.0
django-storages[google]==1.14.2
